requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML dependencies
insightface==0.7.3
onnxruntime>=1.17.0
# onnxruntime-gpu==1.16.3  # GPU 지원시 주석 해제

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.4

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Cache (Redis support - optional)
redis==5.0.1
hiredis==2.2.3

# Monitoring (optional)
prometheus-client==0.19.0

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
# pre-commit==3.6.0